Fix some Corpse typechecking issues
Still need to fix problem where animation might not sync up with the
clock, leading to the first frame of the death animation getting
displayed right before the Corpse despawns. I suspect this is because of
float error between the clock (whose ULP keeps growing) differs from the
Anim's (whose ULP is always <= epsilon/2).

Might need to create a different Anim class which links to a clock on an
Entity, and provide a lazy clock on all entities so it's always there.